feat: 내 앨범 페이지 get 통신 및 출력 완료

+++ src/components/SelectedContentRemoveBtn.jsx
@@ -5,10 +5,14 @@ import styled from 'styled-components';
 function SelectedContentRemoveBtn({ setState, selectedContent, __removeContent }) {
     const dispatch = useDispatch();
     const removeSelectedContent = () => {
-        selectedContent.forEach((contentId) => {
-            dispatch(__removeContent(contentId));
-        });
-        setState({ editClicked: false, selectedContent: [] });
+        if (window.confirm("정말로 삭제 하시겠습니까? 확인시 모든 데이터가 삭제 됩니다.")) {
+            selectedContent.forEach((contentId) => {
+                dispatch(__removeContent(contentId));
+            });
+            setState({ editClicked: false, selectedContent: [] })
+        } else {
+
+        };
     };
 
     return (
